Backend/requirements.txt
fastapi
uvicorn[standard]
python-socketio[asyncio]
psycopg2-binary
python-dotenv
starlette
pydantic[email]
python-jose
passlib[bcrypt]
asyncpg
sqlalchemy
alembic
PyJWT
bcrypt
cloudinary